requirements.txt
uvicorn~=0.34.2
fastapi~=0.115.12
pillow~=11.2.1
toml~=0.10.2